#250: Reorganize related unit tests.
